+++ packages/react/app/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Change Log
 
+## 9.0.0-msal-v5.1
+
+### Patch Changes
+
+- [#3944](https://github.com/equinor/fusion-framework/pull/3944) [`312755f`](https://github.com/equinor/fusion-framework/commit/312755f01c7592329aec847ee4956fe9bf58458f) Thanks [@dependabot](https://github.com/apps/dependabot)! - pre-release msal v5
+
+- Updated dependencies [[`312755f`](https://github.com/equinor/fusion-framework/commit/312755f01c7592329aec847ee4956fe9bf58458f)]:
+  - @equinor/fusion-framework-module-msal@7.0.0-msal-v5.1
+  - @equinor/fusion-framework-app@10.2.2-msal-v5.1
+  - @equinor/fusion-framework-module-app@7.2.3-msal-v5.0
+  - @equinor/fusion-framework-module@5.0.6-msal-v5.0
+  - @equinor/fusion-framework-module-navigation@6.0.1-msal-v5.0
+  - @equinor/fusion-framework-react@7.4.20-msal-v5.1
+  - @equinor/fusion-framework-react-module-http@10.0.1-msal-v5.1
+  - @equinor/fusion-framework-react-module@3.1.14-msal-v5.0
+
+## 9.0.0-msal-v5.0
+
+### Patch Changes
+
+- Updated dependencies [[`f373ec5`](https://github.com/equinor/fusion-framework/commit/f373ec5573d70a1b23e2b6261c831d3b23a8cf98)]:
+  - @equinor/fusion-framework-module-msal@7.0.0-msal-v5.0
+  - @equinor/fusion-framework-app@10.2.2-msal-v5.0
+  - @equinor/fusion-framework-module-app@7.2.2
+  - @equinor/fusion-framework-react@7.4.20-msal-v5.0
+  - @equinor/fusion-framework-react-module-http@10.0.1-msal-v5.0
+
 ## 8.2.0
 
 ### Minor Changes